+++ components/Table/EventTable/ProgramEventTableImpl.tsx
@@ -11,7 +11,7 @@ import CreateEventPopupWindow from '@/components/Forms/EventPopupWindowForm';
 import TagEventPopupWindow from '@/components/Forms/TagEventPopupWindow';
 import { EventDataIndex, EventRowData } from '@/utils/table-types';
 import { fetcher, handleExport } from '@/utils/utils';
-import TableHeader from '@/components/table/event-table/TableHeader';
+import TableHeader from '@/components/Table/EventTable/TableHeader';
 import { convertEventDataToRowData } from '@/utils/table-utils';
 import useSWR from 'swr';
 import { QueriedVolunteerEventDTO } from 'bookem-shared/src/types/database';
